Updated Shift & LibraryHour
Mostly done the backend implementation. Need to do Shift's controller methods. #27 #28

+++ LibrarySystem-Backend/src/main/java/ca/mcgill/ecse321/librarysystem/service/ShiftService.java
@@ -20,9 +20,10 @@ public class ShiftService {
 	LibrarianRepository librarianRepo; 
 
 
-	/* @author Arman
+	/**
+	 * @author Arman
 	 * @param librarianId, startDate, endDate, startTime, endTime
-	 * @return shift
+	 * @return shift 
 	 * Note that there is a start and an end date to account for overnight shifts
 	 */
 	@Transactional 
@@ -48,7 +49,8 @@ public class ShiftService {
 		shiftRepo.save(shift); 
 		return shift;
 	}
-	/* @author Arman
+	/**
+	 * @author Arman
 	 * @param librarianId, oldStartDate, oldStartTime, startDate, endDate, startTime, endTime
 	 * @return shift
 	 */
@@ -78,7 +80,8 @@ public class ShiftService {
 		return shift; 
 	}
 	
-	/* @author Arman
+	/**
+	 * @author Arman
 	 * @param librarianId, startTime, endTime
 	 * @return true if the shift is deleted
 	 */
@@ -87,10 +90,91 @@ public class ShiftService {
 		User user = null;
 
 		if (!(user instanceof HeadLibrarian)) throw new IllegalArgumentException("Only the Head Librarian can modify librarian shifts");
+		if (librarianId==null || startDate==null || startTime==null) {
+			throw new IllegalArgumentException("Fields cannot be blank"); 
+		}
 		Shift shift = shiftRepo.findShiftByLibrarianIdAndStartDateAndStartTime(librarianId, startDate, startTime); 
 		if (shift==null) throw new IllegalArgumentException("Shift cannot be found"); 
 		shiftRepo.delete(shiftRepo.findShiftByLibrarianIdAndStartDateAndStartTime(librarianId, startDate, startTime));
 		return true; 
+	}
+	
+	/**
+	 * @author Arman
+	 * @param librarianId
+	 * @return true if all their shifts are deleted
+	 */
+	@Transactional 
+	public boolean deleteLibrarianShifts(String librarianId) {
+		// check that it's head librarian 
+		if (librarianId == null) throw new IllegalArgumentException("librarian ID cannot be blank"); 
+		// Searches through all the shifts, deleting the ones associated with a certain librarian
+		List<Shift> allShifts = toList(shiftRepo.findAll()); 
+		for (Shift shift: allShifts) {
+			if (shift.getLibrarianId().equals(librarianId)) {
+				shiftRepo.delete(shift);
+			}
+		}
+		return true; 
+	}
+	
+	/**
+	 *@author Arman
+	 *@param librarianId, startTime, endTime
+	 *@return shift
+	 *Note that librarians can view (but not modify) each other's shifts
+	 */
+	@Transactional 
+	public Shift getShift(String librarianId, Date startDate, Time startTime) {
+		// check if user is a librarian 
+		
+		if (librarianId==null || startDate==null || startTime==null) {
+			throw new IllegalArgumentException("Fields cannot be blank"); 
+		}
+		
+		Shift shift = shiftRepo.findShiftByLibrarianIdAndStartDateAndStartTime(librarianId, startDate, startTime); 
+		if (shift==null) throw new IllegalArgumentException("Shift cannot be found"); 
+		return shift; 
+	}
+	/**
+	 *@author Arman
+	 *@param librarianId
+	 *@return list of all the librarian's shifts 
+	 */
+	@Transactional 
+	public List<Shift> getAllShiftsFromLibrarian(String librarianId) {
+		// check if user is a librarian
+		if (librarianId==null) throw new IllegalArgumentException("Librarian ID cannot be blank"); 
+		
+		// This searches through all shifts, add the ones with desired librarianId to a list then return the list. 
+		List<Shift> allShifts = toList(shiftRepo.findAll()); 
+		ArrayList<Shift> desiredShifts = new ArrayList<Shift>(); 
+		for (Shift shift: allShifts) {
+			if (shift.getLibrarianId().equals(librarianId)) {
+				desiredShifts.add(shift); 
+			}
+		}
+		
+		return desiredShifts; 
+	}
+	
+	/**
+	 *@author Arman
+	 *@param none
+	 *@return list of everyone's shifts
+	 */
+	@Transactional 
+	public List<Shift> getAllShifts() {
+		// check if user is a librarian
+		return toList(shiftRepo.findAll()); 
+	}
+	
+	private <T> List<T> toList(Iterable<T> iterable){
+		List<T> resultList = new ArrayList<T>();
+		for (T t : iterable) {
+			resultList.add(t);
+		}
+		return resultList;
 	}
 	
 }
